simplify tui
